refactor: domain에 있던 dto 클래스 view패키지르 이동
domain -> view
dto들이 view의 성향이 강하기 때문에 이동함
