+++ src/main/java/racing/view/dto/CarDto.java
@@ -1,4 +1,4 @@
-package racing.domain.dto;
+package racing.view.dto;
 
 import racing.domain.Car;
